fix: рефакторит utils
