+++ src/components/map/utils.ts
@@ -1,7 +1,28 @@
-import {adaptCityObjectToMap, adaptOffersToMapPoints} from '../../utils';
 import {TCity, TOffer, TOfferPreview} from '../../types/offers';
-import {TMapPoint} from '../../types/map';
+import {TMapCity, TMapPoint} from '../../types/map';
 import {DefaultCity} from '../../const';
+
+export function adaptCityObjectToMap(city: TCity): TMapCity {
+  return {
+    title: city.name,
+    lat: city.location.latitude,
+    lng: city.location.longitude,
+    zoom: city.location.zoom
+  };
+}
+
+export function adaptOfferToMapPoint(offer: TOfferPreview | TOffer): TMapPoint {
+  return {
+    id: offer.id,
+    title: offer.title,
+    lat: offer.location.latitude,
+    lng: offer.location.longitude
+  };
+}
+
+function adaptOffersToMapPoints(offers: Array<TOfferPreview | TOffer>): TMapPoint[] {
+  return offers.map((offer) => adaptOfferToMapPoint(offer));
+}
 
 export default function getMapData(offers: Array<TOfferPreview | TOffer>, defaultActiveId: string | null = null) {
   let city: TCity | undefined = DefaultCity;
